+++ src/components/modules/blockManager.ts
@@ -11,7 +11,7 @@ import Module from '../__module';
 import $ from '../dom';
 import * as _ from '../utils';
 import Blocks from '../blocks';
-import { BlockTool, BlockToolConstructable, BlockToolData, PasteEvent, ToolConfig } from '../../../types';
+import { BlockToolConstructable, BlockToolData, PasteEvent } from '../../../types';
 
 /**
  * @typedef {BlockManager} BlockManager
@@ -204,16 +204,21 @@ export default class BlockManager extends Module {
   /**
    * Creates Block instance by tool name
    *
-   * @param {string} toolName - tools passed in editor config {@link EditorConfig#tools}
-   * @param {object} data - constructor params
-   * @param {object} settings - block settings
+   * @param {string} tool - tools passed in editor config {@link EditorConfig#tools}
+   * @param {BlockToolData} [data] - constructor params
    *
    * @returns {Block}
    */
-  public composeBlock(toolName: string, data: BlockToolData = {}, settings: ToolConfig = {}): Block {
-    const toolInstance = this.Editor.Tools.construct(toolName, data) as BlockTool;
-    const toolClass = this.Editor.Tools.available[toolName] as BlockToolConstructable;
-    const block = new Block(toolName, toolInstance, toolClass, settings, this.Editor.API);
+  public composeBlock({ tool, data = {} }: {tool: string; data?: BlockToolData}): Block {
+    const settings = this.Editor.Tools.getToolSettings(tool);
+    const Tool = this.Editor.Tools.available[tool] as BlockToolConstructable;
+    const block = new Block({
+      name: tool,
+      data,
+      Tool,
+      settings,
+      api: this.Editor.API,
+    });
 
     this.bindEvents(block);
 
@@ -223,30 +228,61 @@ export default class BlockManager extends Module {
   /**
    * Insert new block into _blocks
    *
-   * @param {string} toolName — plugin name, by default method inserts initial block type
+   * @param {string} tool — plugin name, by default method inserts initial block type
    * @param {object} data — plugin data
-   * @param {object} settings - default settings
    * @param {number} index - index where to insert new Block
    * @param {boolean} needToFocus - flag shows if needed to update current Block index
+   * @param {boolean} replace - flag shows if block by passed index should be replaced with inserted one
    *
    * @returns {Block}
    */
-  public insert(
-    toolName: string = this.config.initialBlock,
-    data: BlockToolData = {},
-    settings: ToolConfig = {},
-    index: number = this.currentBlockIndex + 1,
-    needToFocus = true
-  ): Block {
-    const block = this.composeBlock(toolName, data, settings);
+  public insert({
+    tool = this.config.initialBlock,
+    data = {},
+    index = this.currentBlockIndex + 1,
+    needToFocus = true,
+    replace = false,
+  }: {
+    tool?: string;
+    data?: BlockToolData;
+    index?: number;
+    needToFocus?: boolean;
+    replace?: boolean;
+  } = {}): Block {
+    const block = this.composeBlock({
+      tool,
+      data,
+    });
 
-    this._blocks[index] = block;
+    this._blocks.insert(index, block, replace);
 
     if (needToFocus) {
       this.currentBlockIndex = index;
+    } else if (index <= this.currentBlockIndex) {
+      this.currentBlockIndex++;
     }
 
     return block;
+  }
+
+  /**
+   * Replace current working block
+   *
+   * @param {string} tool — plugin name
+   * @param {BlockToolData} data — plugin data
+   *
+   * @returns {Block}
+   */
+  public replace({
+    tool = this.config.initialBlock,
+    data = {},
+  }): Block {
+    return this.insert({
+      tool,
+      data,
+      index: this.currentBlockIndex,
+      replace: true,
+    });
   }
 
   /**
@@ -261,13 +297,10 @@ export default class BlockManager extends Module {
     pasteEvent: PasteEvent,
     replace = false
   ): Block {
-    let block;
-
-    if (replace) {
-      block = this.replace(toolName);
-    } else {
-      block = this.insert(toolName);
-    }
+    const block = this.insert({
+      tool: toolName,
+      replace,
+    });
 
     try {
       block.call(BlockToolAPI.ON_PASTE, pasteEvent);
@@ -289,7 +322,7 @@ export default class BlockManager extends Module {
    * @returns {Block} inserted Block
    */
   public insertInitialBlockAtIndex(index: number, needToFocus = false): Block {
-    const block = this.composeBlock(this.config.initialBlock, {}, {});
+    const block = this.composeBlock({ tool: this.config.initialBlock });
 
     this._blocks[index] = block;
 
@@ -381,7 +414,7 @@ export default class BlockManager extends Module {
    *
    * @returns {number|undefined}
    */
-  public removeSelectedBlocks(): number|undefined {
+  public removeSelectedBlocks(): number | undefined {
     let firstSelectedBlockIndex;
 
     /**
@@ -439,28 +472,7 @@ export default class BlockManager extends Module {
      *
      * @type {Block}
      */
-    return this.insert(this.config.initialBlock, data);
-  }
-
-  /**
-   * Replace current working block
-   *
-   * @param {string} toolName — plugin name
-   * @param {BlockToolData} data — plugin data
-   * @param {ToolConfig} settings — plugin config
-   *
-   * @returns {Block}
-   */
-  public replace(
-    toolName: string = this.config.initialBlock,
-    data: BlockToolData = {},
-    settings: ToolConfig = {}
-  ): Block {
-    const block = this.composeBlock(toolName, data, settings);
-
-    this._blocks.insert(this.currentBlockIndex, block, true);
-
-    return block;
+    return this.insert({ data });
   }
 
   /**
@@ -637,7 +649,7 @@ export default class BlockManager extends Module {
     this.dropPointer();
 
     if (needAddInitialBlock) {
-      this.insert(this.config.initialBlock);
+      this.insert();
     }
 
     /**