+++ src/components/tools/stub/index.ts
@@ -1,5 +1,5 @@
 import $ from '../../dom';
-import { API, BlockTool, BlockToolData } from '../../../../types';
+import { API, BlockTool, BlockToolData, BlockToolConstructorOptions } from '../../../../types';
 
 export interface StubData extends BlockToolData{
   title: string;
@@ -52,7 +52,7 @@ export default class Stub implements BlockTool {
    * @param data - stub tool data
    * @param api - Editor.js API
    */
-  constructor({ data, api }: {data: StubData; api: API}) {
+  constructor({ data, api }: BlockToolConstructorOptions<StubData>) {
     this.api = api;
     this.title = data.title || this.api.i18n.t('Error');
     this.subtitle = this.api.i18n.t('The block can not be displayed correctly.');